requirements.txt
Django=4.1
django-quill-editor==0.1.40
django-simple-captcha==0.5.17
Pillow==9.2.0
